move def-id to rustc crate
